Create Maze_Solver.py
A basic AI model that uses Q-learning, which is a type of reinforcement learning, to solve a provided maze. 
It takes in a maze in the form of nested arrays, where...
"S" represents the start point
"G" represents the goal/end point
"#" represents a wall(unusable)
" " represents empty space(usable)

The start and goal coordinates are needed to be inputted manually.